refactor mock tasklist data pass from parent
